refactor(ui): dedup HTTP/signed-terms/hex helpers into shared papi package
s3-client, drive-client and the test-helpers buckets module each carried their
own byte-identical copies of the same plumbing. Consolidate into
@web3-storage/papi:

- httpFetch (+ sleep / isAbortError / isRetryableHttpError retry helpers) and
  the HTTP_RETRY_* constants — moved once into papi, exported.
- buildSignedTermsArgs + the MULTI_SIGNATURE_VARIANT table — moved once into
  papi; each client lib re-exports buildSignedTermsArgs from the facade.
- hex conversion: drop the hand-rolled hexToBytes/bytesToHex and reuse
  fromHex/toHex from @polkadot-api/utils (re-exported from papi for a single
  import site). buildSignedTermsArgs now builds the 0x sig payload via toHex,
  and the S3 auth header via toHex — the provider already accepts a 0x prefix
  (it strips it; its own test sends 0x), so the wire format is unchanged.

Net -181 lines across the three modules.

Verified on toaster: tsc -b && vite build pass for drive-ui and s3-ui; test-helpers buckets.ts typechecks clean; drive-ui vitest 18/18 pass; pnpm install --frozen-lockfile accepts the lockfile.
