+++ user-interfaces/s3-ui/src/lib/s3-client.ts
@@ -10,19 +10,20 @@ import { Binary, Enum, type PolkadotSigner, type Transaction, type TxFinalizedPa
 import { Subscription } from "rxjs";
 import { parachain } from "@polkadot-api/descriptors";
 import {
+  buildSignedTermsArgs,
+  httpFetch,
   resolveProviderEndpoint,
+  toHex,
   toSs58,
   type ParachainApi,
   type SignedTerms,
 } from "@web3-storage/papi";
 
-// Re-exported so state modules can keep importing it from the client facade.
+// Re-exported so other modules can keep importing them from the client facade.
+export { buildSignedTermsArgs } from "@web3-storage/papi";
 export type { SignedTerms } from "@web3-storage/papi";
 
 export type Signer = PolkadotSigner;
-
-const HTTP_RETRY_ATTEMPTS = 3;
-const HTTP_RETRY_BASE_MS = 250;
 
 // ── Types ─────────────────────────────────────────────────────────────────────
 
@@ -141,102 +142,6 @@ export interface QueryMatchingProvidersParams {
 }
 
 // ── Helpers ───────────────────────────────────────────────────────────────────
-
-function sleep(ms: number): Promise<void> {
-  return new Promise((r) => setTimeout(r, ms));
-}
-
-function isAbortError(err: unknown): boolean {
-  return (
-    err instanceof DOMException &&
-    (err.name === "AbortError" || err.code === DOMException.ABORT_ERR)
-  );
-}
-
-function isRetryableHttpError(status: number | null): boolean {
-  if (status === null) return true;
-  return status >= 500 && status < 600;
-}
-
-async function httpFetch(
-  url: string,
-  init: RequestInit & { signal?: AbortSignal } = {},
-): Promise<Response> {
-  let lastError: unknown = null;
-  for (let attempt = 0; attempt < HTTP_RETRY_ATTEMPTS; attempt++) {
-    try {
-      const res = await fetch(url, init);
-      if (res.ok || !isRetryableHttpError(res.status)) return res;
-      lastError = new Error(`HTTP ${res.status}: ${await res.text().catch(() => "")}`);
-    } catch (err) {
-      if (isAbortError(err)) throw err;
-      lastError = err;
-    }
-    if (attempt < HTTP_RETRY_ATTEMPTS - 1) {
-      await sleep(HTTP_RETRY_BASE_MS * Math.pow(2, attempt));
-    }
-  }
-  throw lastError instanceof Error ? lastError : new Error("HTTP request failed");
-}
-
-function hexToBytes(hex: string): Uint8Array {
-  const h = hex.startsWith("0x") ? hex.slice(2) : hex;
-  const out = new Uint8Array(h.length / 2);
-  for (let i = 0; i < out.length; i++) {
-    out[i] = parseInt(h.substring(i * 2, i * 2 + 2), 16);
-  }
-  return out;
-}
-
-function bytesToHex(bytes: Uint8Array): string {
-  return Array.from(bytes)
-    .map((b) => b.toString(16).padStart(2, "0"))
-    .join("");
-}
-
-// MultiSignature SCALE variant order from sp_runtime.
-const MULTI_SIGNATURE_VARIANT: Record<number, string> = {
-  0: "Ed25519",
-  1: "Sr25519",
-  2: "Ecdsa",
-  3: "Eth",
-};
-
-export function buildSignedTermsArgs(
-  providerAccount: string,
-  signed: SignedTerms,
-) {
-  const sigBytes = hexToBytes(signed.signature);
-  if (sigBytes.length < 1) {
-    throw new Error("signature too short to contain a MultiSignature variant byte");
-  }
-  const variantByte = sigBytes[0]!;
-  const variantName = MULTI_SIGNATURE_VARIANT[variantByte];
-  if (!variantName) {
-    throw new Error(`unknown MultiSignature variant byte: ${variantByte}`);
-  }
-  const sigPayloadHex =
-    "0x" +
-    Array.from(sigBytes.slice(1))
-      .map((b) => b.toString(16).padStart(2, "0"))
-      .join("");
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const sig = Enum(variantName as any, sigPayloadHex);
-
-  const t = signed.terms;
-  const terms = {
-    owner: t.owner,
-    max_bytes: BigInt(t.max_bytes),
-    duration: t.duration,
-    price_per_byte: BigInt(t.price_per_byte),
-    valid_until: t.valid_until,
-    nonce: BigInt(t.nonce),
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    replica_params: (t.replica_params ?? undefined) as any,
-    bucket_id: t.bucket_id ? BigInt(t.bucket_id) : undefined,
-  };
-  return { provider: providerAccount, terms, sig };
-}
 
 function decodeName(name: unknown): string {
   if (name == null) return "";
@@ -405,8 +310,8 @@ export class S3Client {
     const message = `web3storage:${method}:${Number(bucketId)}:${timestamp}`;
     const msgBytes = new TextEncoder().encode(message);
     const sig = this.keypair.sign(msgBytes);
-    const pubHex = bytesToHex(this.keypair.publicKey);
-    const sigHex = bytesToHex(sig);
+    const pubHex = toHex(this.keypair.publicKey);
+    const sigHex = toHex(sig);
     return { Authorization: `Web3Storage ${pubHex}:${sigHex}:${timestamp}` };
   }
