+++ user-interfaces/shared/papi/package.json
@@ -19,6 +19,7 @@
   "dependencies": {
     "@multiformats/multiaddr-to-uri": "^12.0.0",
     "@polkadot-api/descriptors": "file:.papi/descriptors",
+    "@polkadot-api/utils": "^0.1.0",
     "@polkadot-labs/hdkd-helpers": "^0.0.30",
     "polkadot-api": "^2.1.6"
   }
