+++ user-interfaces/shared/papi/src/index.ts
@@ -7,10 +7,18 @@
 import { ss58Address, ss58Decode } from '@polkadot-labs/hdkd-helpers'
 import { multiaddrToUri } from "@multiformats/multiaddr-to-uri";
 import { parachain } from "@polkadot-api/descriptors";
-import type { TypedApi } from "polkadot-api";
+import { Enum, type TypedApi } from "polkadot-api";
+import { fromHex, toHex } from "@polkadot-api/utils";
 
 /** Typed parachain API, shared by every UI that talks to the chain. */
 export type ParachainApi = TypedApi<typeof parachain>;
+
+/**
+ * Hex helpers re-exported from `@polkadot-api/utils` so UIs have one import
+ * site: `toHex` emits a `0x`-prefixed string, `fromHex` accepts hex with or
+ * without the prefix.
+ */
+export { fromHex, toHex };
 
 /**
  * SS58 address prefix used to encode public keys for display.
@@ -164,6 +172,95 @@ export async function negotiateProviderTerms(
       error: err instanceof Error ? err.message : "Failed to negotiate with provider",
     };
   }
+}
+
+const HTTP_RETRY_ATTEMPTS = 3;
+const HTTP_RETRY_BASE_MS = 250;
+
+function sleep(ms: number): Promise<void> {
+  return new Promise((r) => setTimeout(r, ms));
+}
+
+function isAbortError(err: unknown): boolean {
+  return (
+    err instanceof DOMException &&
+    (err.name === "AbortError" || err.code === DOMException.ABORT_ERR)
+  );
+}
+
+function isRetryableHttpError(status: number | null): boolean {
+  if (status === null) return true;
+  return status >= 500 && status < 600;
+}
+
+/**
+ * `fetch` with bounded exponential-backoff retry on transient failures (5xx and
+ * network errors). Aborts (via `init.signal`) propagate immediately and are
+ * never retried.
+ */
+export async function httpFetch(
+  url: string,
+  init: RequestInit & { signal?: AbortSignal } = {},
+): Promise<Response> {
+  let lastError: unknown = null;
+  for (let attempt = 0; attempt < HTTP_RETRY_ATTEMPTS; attempt++) {
+    try {
+      const res = await fetch(url, init);
+      if (res.ok || !isRetryableHttpError(res.status)) return res;
+      lastError = new Error(`HTTP ${res.status}: ${await res.text().catch(() => "")}`);
+    } catch (err) {
+      if (isAbortError(err)) throw err;
+      lastError = err;
+    }
+    if (attempt < HTTP_RETRY_ATTEMPTS - 1) {
+      await sleep(HTTP_RETRY_BASE_MS * Math.pow(2, attempt));
+    }
+  }
+  throw lastError instanceof Error ? lastError : new Error("HTTP request failed");
+}
+
+// SCALE variant ordering of sp_runtime::MultiSignature.
+const MULTI_SIGNATURE_VARIANT: Record<number, string> = {
+  0: "Ed25519",
+  1: "Sr25519",
+  2: "Ecdsa",
+  3: "Eth",
+};
+
+/**
+ * Build the `{ provider, terms, sig }` args shared by every signed-terms
+ * extrinsic.
+ *
+ * The inner of `MultiSignature::Sr25519` is `[u8; 64]`, which PAPI v2 encodes
+ * as `SizedBytes(64) = Codec<string>` — so the signature payload is passed as a
+ * `0x`-prefixed hex string (via `toHex`), not a `Uint8Array`.
+ */
+export function buildSignedTermsArgs(providerAccount: string, signed: SignedTerms) {
+  const sigBytes = fromHex(signed.signature);
+  if (sigBytes.length < 1) {
+    throw new Error("signature too short to contain a MultiSignature variant byte");
+  }
+  const variantByte = sigBytes[0]!;
+  const variantName = MULTI_SIGNATURE_VARIANT[variantByte];
+  if (!variantName) {
+    throw new Error(`unknown MultiSignature variant byte: ${variantByte}`);
+  }
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  const sig = Enum(variantName as any, toHex(sigBytes.slice(1)));
+
+  const t = signed.terms;
+  const terms = {
+    owner: t.owner,
+    max_bytes: BigInt(t.max_bytes),
+    duration: t.duration,
+    price_per_byte: BigInt(t.price_per_byte),
+    valid_until: t.valid_until,
+    nonce: BigInt(t.nonce),
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    replica_params: (t.replica_params ?? undefined) as any,
+    bucket_id: t.bucket_id ? BigInt(t.bucket_id) : undefined,
+  };
+  return { provider: providerAccount, terms, sig };
 }
 
 /**
