+++ user-interfaces/shared/test-helpers/src/buckets.ts
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
-import { Binary, Enum } from "polkadot-api";
+import { Binary } from "polkadot-api";
+import { buildSignedTermsArgs, negotiateTerms } from "@web3-storage/papi";
 import { getApi, submitExtrinsic, submitExtrinsicBestBlock } from "./chain-api";
 import { Alice, type DevSigner } from "./signers";
 
@@ -8,105 +9,6 @@ import { Alice, type DevSigner } from "./signers";
 
 const DEFAULT_PROVIDER_URL = "http://127.0.0.1:3333";
 const DEFAULT_PROVIDER_ACCOUNT = Alice.address;
-
-// SCALE variant ordering of sp_runtime::MultiSignature.
-const MULTI_SIGNATURE_VARIANT: Record<number, string> = {
-  0: "Ed25519",
-  1: "Sr25519",
-  2: "Ecdsa",
-  3: "Eth",
-};
-
-interface NegotiateRequest {
-  owner: string;
-  max_bytes: number | bigint;
-  duration: number;
-  price_per_byte: number | bigint;
-  replica_params: unknown | null;
-  bucket_id: bigint | null;
-}
-
-interface SignedTerms {
-  terms: {
-    owner: string;
-    max_bytes: number | bigint;
-    duration: number;
-    price_per_byte: number | bigint;
-    valid_until: number;
-    nonce: number | bigint;
-    replica_params: unknown | null;
-    bucket_id: bigint | null;
-  };
-  signature: string;
-}
-
-async function negotiateTerms(
-  providerUrl: string,
-  request: NegotiateRequest,
-): Promise<SignedTerms> {
-  const res = await fetch(`${providerUrl.replace(/\/$/, "")}/negotiate`, {
-    method: "POST",
-    headers: { "content-type": "application/json" },
-    body: JSON.stringify(request, (_k, v) =>
-      typeof v === "bigint" ? v.toString() : v,
-    ),
-  });
-  if (!res.ok) {
-    const body = await res.text().catch(() => "");
-    throw new Error(`/negotiate failed: ${res.status} ${body}`);
-  }
-  return res.json();
-}
-
-function hexToBytes(hex: string): Uint8Array {
-  const h = hex.startsWith("0x") ? hex.slice(2) : hex;
-  const out = new Uint8Array(h.length / 2);
-  for (let i = 0; i < out.length; i++) {
-    out[i] = parseInt(h.substring(i * 2, i * 2 + 2), 16);
-  }
-  return out;
-}
-
-/**
- * Build the `{ provider, terms, sig }` args shared by every signed-terms
- * extrinsic. Mirrors `console-ui/src/lib/storage.ts::buildSignedTermsArgs`.
- *
- * The inner of `MultiSignature::Sr25519` is `[u8; 64]` which PAPI v2 encodes
- * as `SizedBytes(64) = Codec<string>` — pass a `0x`-prefixed hex string,
- * NOT a `Uint8Array`. (See PAPI v2 migration doc + console-ui's working
- * implementation.)
- */
-function buildSignedTermsArgs(providerAccount: string, signed: SignedTerms) {
-  const sigBytes = hexToBytes(signed.signature);
-  if (sigBytes.length < 1) {
-    throw new Error("signature too short to contain a MultiSignature variant byte");
-  }
-  const variantName = MULTI_SIGNATURE_VARIANT[sigBytes[0]];
-  if (!variantName) {
-    throw new Error(`unknown MultiSignature variant byte: ${sigBytes[0]}`);
-  }
-  const sigPayloadHex =
-    "0x" +
-    Array.from(sigBytes.slice(1))
-      .map((b) => b.toString(16).padStart(2, "0"))
-      .join("");
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const sig = Enum(variantName as any, sigPayloadHex);
-
-  const t = signed.terms;
-  const terms = {
-    owner: t.owner,
-    max_bytes: BigInt(t.max_bytes),
-    duration: t.duration,
-    price_per_byte: BigInt(t.price_per_byte),
-    valid_until: t.valid_until,
-    nonce: BigInt(t.nonce),
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    replica_params: (t.replica_params ?? undefined) as any,
-    bucket_id: t.bucket_id ?? undefined,
-  };
-  return { provider: providerAccount, terms, sig };
-}
 
 // ─── S3 Buckets (console-ui) ─────────────────────────────────────────────────
